# release-2.1: kv: assorted client optimizations