--- a/pkg/server/status/recorder.go
+++ b/pkg/server/status/recorder.go
@@ -536,8 +536,6 @@ func extractValue(mtr interface{}) (float64, error) {
 	switch mtr := mtr.(type) {
 	case float64:
 		return mtr, nil
-	case *metric.CounterWithRates:
-		return float64(mtr.Count()), nil
 	case *metric.Counter:
 		return float64(mtr.Count()), nil
 	case *metric.Gauge: