--- a/pkg/util/metric/registry_test.go
+++ b/pkg/util/metric/registry_test.go
@@ -31,7 +31,6 @@ func (r *Registry) findMetricByName(name string) Iterable {
 // getCounter returns the Counter in this registry with the given name. If a
 // Counter with this name is not present (including if a non-Counter Iterable is
 // registered with the name), nil is returned.
-// For the purpose of this method, a CounterWithRate is a Counter.
 func (r *Registry) getCounter(name string) *Counter {
 	r.Lock()
 	defer r.Unlock()
@@ -43,8 +42,6 @@ func (r *Registry) getCounter(name string) *Counter {
 	switch t := iterable.(type) {
 	case *Counter:
 		return t
-	case *CounterWithRates:
-		return t.Counter
 	default:
 	}
 	return nil
@@ -78,18 +75,14 @@ func TestRegistry(t *testing.T) {
 	topCounter := NewCounter(Metadata{Name: "top.counter"})
 	r.AddMetric(topCounter)
 
-	topCounterWithRates := NewCounterWithRates(Metadata{Name: "top.counterwithrates"})
-	r.AddMetric(topCounterWithRates)
-
 	r.AddMetric(NewHistogram(Metadata{Name: "top.histogram"}, time.Minute, 1000, 3))
 
 	r.AddMetric(NewGauge(Metadata{Name: "bottom.gauge"}))
 	ms := &struct {
-		StructGauge            *Gauge
-		StructGauge64          *GaugeFloat64
-		StructCounter          *Counter
-		StructHistogram        *Histogram
-		StructCounterWithRates *CounterWithRates
+		StructGauge     *Gauge
+		StructGauge64   *GaugeFloat64
+		StructCounter   *Counter
+		StructHistogram *Histogram
 		// A few extra ones: either not exported, or not metric objects.
 		privateStructGauge   *Gauge
 		privateStructGauge64 *GaugeFloat64
@@ -97,31 +90,28 @@ func TestRegistry(t *testing.T) {
 		AlsoNotAMetric       string
 		ReallyNotAMetric     *Registry
 	}{
-		StructGauge:            NewGauge(Metadata{Name: "struct.gauge"}),
-		StructGauge64:          NewGaugeFloat64(Metadata{Name: "struct.gauge64"}),
-		StructCounter:          NewCounter(Metadata{Name: "struct.counter"}),
-		StructHistogram:        NewHistogram(Metadata{Name: "struct.histogram"}, time.Minute, 1000, 3),
-		StructCounterWithRates: NewCounterWithRates(Metadata{Name: "struct.counterwithrates"}),
-		privateStructGauge:     NewGauge(Metadata{Name: "struct.private-gauge"}),
-		privateStructGauge64:   NewGaugeFloat64(Metadata{Name: "struct.private-gauge64"}),
-		NotAMetric:             0,
-		AlsoNotAMetric:         "foo",
-		ReallyNotAMetric:       NewRegistry(),
+		StructGauge:          NewGauge(Metadata{Name: "struct.gauge"}),
+		StructGauge64:        NewGaugeFloat64(Metadata{Name: "struct.gauge64"}),
+		StructCounter:        NewCounter(Metadata{Name: "struct.counter"}),
+		StructHistogram:      NewHistogram(Metadata{Name: "struct.histogram"}, time.Minute, 1000, 3),
+		privateStructGauge:   NewGauge(Metadata{Name: "struct.private-gauge"}),
+		privateStructGauge64: NewGaugeFloat64(Metadata{Name: "struct.private-gauge64"}),
+		NotAMetric:           0,
+		AlsoNotAMetric:       "foo",
+		ReallyNotAMetric:     NewRegistry(),
 	}
 	r.AddMetricStruct(ms)
 
 	expNames := map[string]struct{}{
-		"top.counterwithrates":    {},
-		"top.histogram":           {},
-		"top.gauge":               {},
-		"top.floatgauge":          {},
-		"top.counter":             {},
-		"bottom.gauge":            {},
-		"struct.gauge":            {},
-		"struct.gauge64":          {},
-		"struct.counter":          {},
-		"struct.histogram":        {},
-		"struct.counterwithrates": {},
+		"top.histogram":    {},
+		"top.gauge":        {},
+		"top.floatgauge":   {},
+		"top.counter":      {},
+		"bottom.gauge":     {},
+		"struct.gauge":     {},
+		"struct.gauge64":   {},
+		"struct.counter":   {},
+		"struct.histogram": {},
 	}
 
 	r.Each(func(name string, _ interface{}) {
@@ -147,9 +137,6 @@ func TestRegistry(t *testing.T) {
 
 	if c := r.getCounter("top.counter"); c != topCounter {
 		t.Errorf("getCounter returned %v, expected %v", c, topCounter)
-	}
-	if c, e := r.getCounter("top.counterwithrates"), topCounterWithRates.Counter; c != e {
-		t.Errorf("getCounter returned %v, expected %v", c, e)
 	}
 	if c := r.getCounter("bad"); c != nil {
 		t.Errorf("getCounter returned non-nil %v, expected nil", c)