Backport 3/3 commits from #30448.

/cc @cockroachdb/release

---

See individual commits.

This PR increases throughput on a single node kv-95 workload as follows (measured on a 96-core GCE machine and running the client with concurrency=48).
```
_elapsed___errors_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)__total
  100.0s        0        1257869        12578.5      2.8      3.3      5.5      7.1     48.2  read
  100.0s        0          66469          664.7     19.5     19.9     31.5     35.7     60.8  write
```
to
```
_elapsed___errors_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)__total
  100.0s        0        1369178        13691.4      2.5      2.8      5.2      6.8    503.3  read
  100.0s        0          71850          718.5     18.3     18.9     28.3     33.6    520.1  write
```

The PR also takes away about 3/4 of the performance regression introduced in #28185. See #30208 for more.